+++ buckets.tf
@@ -1,3 +1,8 @@
+locals {
+  buckets = {
+    for item in var.services : item.name => item if item.type == "bucket"
+  }
+}
 resource "google_compute_backend_bucket" "bucket" {
   for_each    = local.buckets
   name        = "${each.key}-${local.random_suffix}"