+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v6.0.0] - 2022-08-21
+## [v6.1.0] - 2023-12-21
+### Added
+- support for custom backends to enable rate balancing
+
+## [v6.0.0] - 2023-11-30
 ### Added
 - support for paths
 - random suffix size to easy migrations